+++ stop_signon.sh
@@ -4,7 +4,7 @@ set -ex
 
 GEM_ROOT=$(cd $(dirname $0) ; pwd)
 TMP_ROOT=${GEM_ROOT}/tmp
-APP_ROOT=${TMP_ROOT}/signonotron2
+APP_ROOT=${TMP_ROOT}/signon
 PID_FILE=${APP_ROOT}/server.pid
 
 if [[ -f ${PID_FILE} && -n $(cat ${PID_FILE}) ]]